(#3281) - refactor websql
Also moved http.js to http/http.js for consistency.

Also finally got rid of the need to manually add ok=true
for `bulkDocs`'s _local responses; `utils.processDocs` now
handles that.
