+++ lib/adapters/http/http.js
@@ -9,8 +9,8 @@ var CHANGES_BATCH_SIZE = 25;
 // TODO: we could measure the full URL to enforce exactly 2000 chars
 var MAX_URL_LENGTH = 1800;
 
-var utils = require('../utils');
-var errors = require('../deps/errors');
+var utils = require('../../utils');
+var errors = require('../../deps/errors');
 var log = require('debug')('pouchdb:http');
 var isBrowser = typeof process === 'undefined' || process.browser;
