+++ lib/adapters/idb/idb-bulk-docs.js
@@ -119,18 +119,8 @@ function idbBulkDocs(req, opts, api, idb, Changes, callback) {
   }
 
   function complete() {
-
     if (preconditionErrored) {
       return;
-    }
-
-    for (var i = 0, len = results.length; i < len; i++) {
-      var result = results[i];
-      if (!result.error) {
-        // TODO: this should be set in utils.processDocs
-        // for local docs
-        result.ok = true;
-      }
     }
 
     Changes.notify(api._name);
