+++ lib/adapters/leveldb/leveldb.js
@@ -604,7 +604,11 @@ function LevelPouch(opts, callback) {
           valueEncoding: safeJsonEncoding
         }];
         txn.batch(batch);
-        results[index] = doc;
+        results[index] = {
+          ok: true,
+          id: doc.metadata.id,
+          rev: winningRev
+        };
         fetchedDocs.set(doc.metadata.id, doc.metadata);
         callback2();
       }
@@ -707,25 +711,6 @@ function LevelPouch(opts, callback) {
           callback(err);
         });
       }
-      var aresults = results.map(function (result) {
-        if (!Object.keys(result).length) {
-          return {
-            ok: true
-          };
-        }
-        if (result.error) {
-          return result;
-        }
-
-        var metadata = result.metadata;
-        var rev = merge.winningRev(metadata);
-
-        return {
-          ok: true,
-          id: metadata.id,
-          rev: rev
-        };
-      });
       txn.batch([
         {
           prefix: stores.metaStore,
@@ -750,7 +735,7 @@ function LevelPouch(opts, callback) {
         db._updateSeq = newUpdateSeq;
         LevelPouch.Changes.notify(name);
         process.nextTick(function () {
-          callback(null, aresults);
+          callback(null, results);
         });
       });
     }
