+++ lib/adapters/websql/websql.js
@@ -1,44 +1,29 @@
 'use strict';
 
-var utils = require('../utils');
-var merge = require('../merge');
-var errors = require('../deps/errors');
-var parseHexString = require('../deps/parse-hex');
+var utils = require('../../utils');
+var merge = require('../../merge');
+var errors = require('../../deps/errors');
+var parseHexString = require('../../deps/parse-hex');
 
-function quote(str) {
-  return "'" + str + "'";
-}
+var websqlConstants = require('./websql-constants');
+var websqlUtils = require('./websql-utils');
+var websqlBulkDocs = require('./websql-bulk-docs');
 
-// escapeBlob and unescapeBlob are workarounds for a websql bug:
-// https://code.google.com/p/chromium/issues/detail?id=422690
-// https://bugs.webkit.org/show_bug.cgi?id=137637
-// The goal is to never actually insert the \u0000 character
-// in the database.
-function escapeBlob(str) {
-  return str
-    .replace(/\u0002/g, '\u0002\u0002')
-    .replace(/\u0001/g, '\u0001\u0002')
-    .replace(/\u0000/g, '\u0001\u0001');
-}
+var ADAPTER_VERSION = websqlConstants.ADAPTER_VERSION;
+var DOC_STORE = websqlConstants.DOC_STORE;
+var BY_SEQ_STORE = websqlConstants.BY_SEQ_STORE;
+var ATTACH_STORE = websqlConstants.ATTACH_STORE;
+var LOCAL_STORE = websqlConstants.LOCAL_STORE;
+var META_STORE = websqlConstants.META_STORE;
+var ATTACH_AND_SEQ_STORE = websqlConstants.ATTACH_AND_SEQ_STORE;
 
-function unescapeBlob(str) {
-  return str
-    .replace(/\u0001\u0001/g, '\u0000')
-    .replace(/\u0001\u0002/g, '\u0001')
-    .replace(/\u0002\u0002/g, '\u0002');
-}
-
-// question mark groups IN queries, e.g. 3 -> '(?,?,?)'
-function qMarks(num) {
-  var s = '(';
-  while (num--) {
-    s += '?';
-    if (num) {
-      s += ',';
-    }
-  }
-  return s + ')';
-}
+var qMarks = websqlUtils.qMarks;
+var stringifyDoc = websqlUtils.stringifyDoc;
+var unstringifyDoc = websqlUtils.unstringifyDoc;
+var select = websqlUtils.select;
+var compactRevs = websqlUtils.compactRevs;
+var unknownError = websqlUtils.unknownError;
+var getSize = websqlUtils.getSize;
 
 function fetchAttachmentsIfNecessary(doc, opts, api, txn, cb) {
   var attachments = Object.keys(doc._attachments || {});
@@ -75,91 +60,6 @@ function fetchAttachmentsIfNecessary(doc, opts, api, txn, cb) {
   });
 }
 
-function compactRevs(revs, docId, tx) {
-
-  if (!revs.length) {
-    return;
-  }
-
-  var numDone = 0;
-  var seqs = [];
-
-  function checkDone() {
-    if (++numDone === revs.length) { // done
-      deleteOrphans();
-    }
-  }
-
-  function deleteOrphans() {
-    // find orphaned attachment digests
-
-    if (!seqs.length) {
-      return;
-    }
-
-    var sql = 'SELECT DISTINCT digest AS digest FROM ' +
-      ATTACH_AND_SEQ_STORE + ' WHERE seq IN (' +
-        seqs.map(function () { return '?'; }).join(',') +
-        ');';
-
-    tx.executeSql(sql, seqs, function (tx, res) {
-
-      var digestsToCheck = [];
-      for (var i = 0; i < res.rows.length; i++) {
-        digestsToCheck.push(res.rows.item(i).digest);
-      }
-      if (!digestsToCheck.length) {
-        return;
-      }
-
-      var sql = 'DELETE FROM ' + ATTACH_AND_SEQ_STORE +
-        ' WHERE seq IN (' +
-        seqs.map(function () { return '?'; }).join(',') +
-        ')';
-      tx.executeSql(sql, seqs, function (tx) {
-
-        var sql = 'SELECT digest FROM ' + ATTACH_AND_SEQ_STORE +
-          ' WHERE digest IN (' +
-          digestsToCheck.map(function () { return '?'; }).join(',') +
-          ')';
-        tx.executeSql(sql, digestsToCheck, function (tx, res) {
-          var nonOrphanedDigests = new utils.Set();
-          for (var i = 0; i < res.rows.length; i++) {
-            nonOrphanedDigests.add(res.rows.item(i).digest);
-          }
-          digestsToCheck.forEach(function (digest) {
-            if (nonOrphanedDigests.has(digest)) {
-              return;
-            }
-            tx.executeSql(
-              'DELETE FROM ' + ATTACH_AND_SEQ_STORE + ' WHERE digest=?',
-              [digest]);
-            tx.executeSql(
-              'DELETE FROM ' + ATTACH_STORE + ' WHERE digest=?', [digest]);
-          });
-        });
-      });
-    });
-  }
-
-    // update by-seq and attach stores in parallel
-  revs.forEach(function (rev) {
-    var sql = 'SELECT seq FROM ' + BY_SEQ_STORE +
-      ' WHERE doc_id=? AND rev=?';
-
-    tx.executeSql(sql, [docId, rev], function (tx, res) {
-      if (!res.rows.length) { // already deleted
-        return checkDone();
-      }
-      var seq = res.rows.item(0).seq;
-      seqs.push(seq);
-
-      tx.executeSql(
-        'DELETE FROM ' + BY_SEQ_STORE + ' WHERE seq=?', [seq], checkDone);
-    });
-  });
-}
-
 var cachedDatabases = {};
 
 var openDBFunction = (typeof navigator !== 'undefined' &&
@@ -181,21 +81,6 @@ function openDB(name, version, desc, size) {
 }
 
 var POUCH_VERSION = 1;
-var ADAPTER_VERSION = 7; // used to manage migrations
-
-// The object stores created for each database
-// DOC_STORE stores the document meta data, its revision history and state
-var DOC_STORE = quote('document-store');
-// BY_SEQ_STORE stores a particular version of a document, keyed by its
-// sequence id
-var BY_SEQ_STORE = quote('by-sequence');
-// Where we store attachments
-var ATTACH_STORE = quote('attach-store');
-var LOCAL_STORE = quote('local-store');
-var META_STORE = quote('metadata-store');
-// where we store many-to-many relations between attachment
-// digests and seqs
-var ATTACH_AND_SEQ_STORE = quote('attach-seq-store');
 
 // these indexes cover the ground for most allDocs queries
 var BY_SEQ_STORE_DELETED_INDEX_SQL =
@@ -223,67 +108,17 @@ var SELECT_DOCS = BY_SEQ_STORE + '.seq AS seq, ' +
   BY_SEQ_STORE + '.rev AS rev, ' +
   DOC_STORE + '.json AS metadata';
 
-function select(selector, table, joiner, where, orderBy) {
-  return 'SELECT ' + selector + ' FROM ' +
-    (typeof table === 'string' ? table : table.join(' JOIN ')) +
-    (joiner ? (' ON ' + joiner) : '') +
-    (where ? (' WHERE ' +
-      (typeof where === 'string' ? where : where.join(' AND '))) : '') +
-    (orderBy ? (' ORDER BY ' + orderBy) : '');
-}
-
-function unknownError(callback) {
-  return function (event) {
-    // event may actually be a SQLError object, so report is as such
-    var errorNameMatch = event && event.constructor.toString()
-      .match(/function ([^\(]+)/);
-    var errorName = (errorNameMatch && errorNameMatch[1]) || event.type;
-    var errorReason = event.target || event.message;
-    callback(errors.error(errors.WSQ_ERROR, errorReason, errorName));
-  };
-}
-
-function stringifyDoc(doc) {
-  // don't bother storing the id/rev. it uses lots of space,
-  // in persistent map/reduce especially
-  delete doc._id;
-  delete doc._rev;
-  return JSON.stringify(doc);
-}
-
-function unstringifyDoc(doc, id, rev) {
-  doc = JSON.parse(doc);
-  doc._id = id;
-  doc._rev = rev;
-  return doc;
-}
-
-function getSize(opts) {
-  if ('size' in opts) {
-    // triggers immediate popup in iOS, fixes #2347
-    // e.g. 5000001 asks for 5 MB, 10000001 asks for 10 MB,
-    return opts.size * 1000000;
-  }
-  // In iOS, doesn't matter as long as it's <= 5000000.
-  // Except that if you request too much, our tests fail
-  // because of the native "do you accept?" popup.
-  // In Android <=4.3, this value is actually used as an
-  // honest-to-god ceiling for data, so we need to
-  // set it to a decently high number.
-  var isAndroid = /Android/.test(window.navigator.userAgent);
-  return isAndroid ? 5000000 : 1; // in PhantomJS, if you use 0 it will crash
-}
-
 function WebSqlPouch(opts, callback) {
   var api = this;
   var instanceId = null;
-  var name = opts.name;
   var size = getSize(opts);
   var idRequests = [];
-  var docCount = -1; // cache sqlite count(*) for performance
   var encoding;
 
-  var db = openDB(name, POUCH_VERSION, name, size);
+  api._docCount = -1; // cache sqlite count(*) for performance
+  api._name = opts.name;
+
+  var db = openDB(api._name, POUCH_VERSION, api._name, size);
   if (!db) {
     return callback(errors.error(errors.UNKNOWN_ERROR));
   } else if (typeof db.readTransaction !== 'function') {
@@ -294,7 +129,7 @@ function WebSqlPouch(opts, callback) {
   function dbCreated() {
     // note the db name in case the browser upgrades to idb
     if (utils.hasLocalStorage()) {
-      window.localStorage['_pouch__websqldb_' + name] = true;
+      window.localStorage['_pouch__websqldb_' + api._name] = true;
     }
     callback(null, api);
   }
@@ -673,8 +508,8 @@ function WebSqlPouch(opts, callback) {
 
   if (utils.isCordova()) {
     //to wait until custom api is made in pouch.adapters before doing setup
-    window.addEventListener(name + '_pouch', function cordova_init() {
-      window.removeEventListener(name + '_pouch', cordova_init, false);
+    window.addEventListener(api._name + '_pouch', function cordova_init() {
+      window.removeEventListener(api._name + '_pouch', cordova_init, false);
       setup();
     }, false);
   } else {
@@ -705,324 +540,7 @@ function WebSqlPouch(opts, callback) {
   };
 
   api._bulkDocs = function (req, opts, callback) {
-
-    var newEdits = opts.new_edits;
-    var userDocs = req.docs;
-
-    // Parse the docs, give them a sequence number for the result
-    var docInfos = userDocs.map(function (doc, i) {
-      if (doc._id && utils.isLocalId(doc._id)) {
-        return doc;
-      }
-      var newDoc = utils.parseDoc(doc, newEdits);
-      return newDoc;
-    });
-
-    var docInfoErrors = docInfos.filter(function (docInfo) {
-      return docInfo.error;
-    });
-    if (docInfoErrors.length) {
-      return callback(docInfoErrors[0]);
-    }
-
-    var tx;
-    var results = new Array(docInfos.length);
-    var fetchedDocs = new utils.Map();
-
-    var preconditionErrored;
-    function complete() {
-      if (preconditionErrored) {
-        return callback(preconditionErrored);
-      }
-      var aresults = results.map(function (result) {
-        if (!Object.keys(result).length) {
-          return {
-            ok: true
-          };
-        }
-        if (result.error) {
-          return result;
-        }
-
-        var metadata = result.metadata;
-        var rev = merge.winningRev(metadata);
-
-        return {
-          ok: true,
-          id: metadata.id,
-          rev: rev
-        };
-      });
-      WebSqlPouch.Changes.notify(name);
-      docCount = -1; // invalidate
-      callback(null, aresults);
-    }
-
-    function verifyAttachment(digest, callback) {
-      var sql = 'SELECT count(*) as cnt FROM ' + ATTACH_STORE +
-        ' WHERE digest=?';
-      tx.executeSql(sql, [digest], function (tx, result) {
-        if (result.rows.item(0).cnt === 0) {
-          var err = errors.error(errors.MISSING_STUB,
-                                'unknown stub attachment with digest ' +
-                                digest);
-          callback(err);
-        } else {
-          callback();
-        }
-      });
-    }
-
-    function verifyAttachments(finish) {
-      var digests = [];
-      docInfos.forEach(function (docInfo) {
-        if (docInfo.data && docInfo.data._attachments) {
-          Object.keys(docInfo.data._attachments).forEach(function (filename) {
-            var att = docInfo.data._attachments[filename];
-            if (att.stub) {
-              digests.push(att.digest);
-            }
-          });
-        }
-      });
-      if (!digests.length) {
-        return finish();
-      }
-      var numDone = 0;
-      var err;
-
-      function checkDone() {
-        if (++numDone === digests.length) {
-          finish(err);
-        }
-      }
-      digests.forEach(function (digest) {
-        verifyAttachment(digest, function (attErr) {
-          if (attErr && !err) {
-            err = attErr;
-          }
-          checkDone();
-        });
-      });
-    }
-
-
-    function writeDoc(docInfo, winningRev, deleted, callback, isUpdate,
-                      docCount, resultsIdx) {
-
-      function finish() {
-        var data = docInfo.data;
-        var deletedInt = deleted ? 1 : 0;
-
-        var id = data._id;
-        var rev = data._rev;
-        var json = stringifyDoc(data);
-        var sql = 'INSERT INTO ' + BY_SEQ_STORE +
-          ' (doc_id, rev, json, deleted) VALUES (?, ?, ?, ?);';
-        var sqlArgs = [id, rev, json, deletedInt];
-
-        // map seqs to attachment digests, which
-        // we will need later during compaction
-        function insertAttachmentMappings(seq, callback) {
-          var attsAdded = 0;
-          var attsToAdd = Object.keys(data._attachments || {});
-
-          if (!attsToAdd.length) {
-            return callback();
-          }
-          function checkDone() {
-            if (++attsAdded === attsToAdd.length) {
-              callback();
-            }
-            return false; // ack handling a constraint error
-          }
-          function add(att) {
-            var sql = 'INSERT INTO ' + ATTACH_AND_SEQ_STORE +
-              ' (digest, seq) VALUES (?,?)';
-            var sqlArgs = [data._attachments[att].digest, seq];
-            tx.executeSql(sql, sqlArgs, checkDone, checkDone);
-            // second callback is for a constaint error, which we ignore
-            // because this docid/rev has already been associated with
-            // the digest (e.g. when new_edits == false)
-          }
-          for (var i = 0; i < attsToAdd.length; i++) {
-            add(attsToAdd[i]); // do in parallel
-          }
-        }
-
-        tx.executeSql(sql, sqlArgs, function (tx, result) {
-          var seq = result.insertId;
-          insertAttachmentMappings(seq, function () {
-            dataWritten(tx, seq);
-          });
-        }, function () {
-          // constraint error, recover by updating instead (see #1638)
-          var fetchSql = select('seq', BY_SEQ_STORE, null,
-            'doc_id=? AND rev=?');
-          tx.executeSql(fetchSql, [id, rev], function (tx, res) {
-            var seq = res.rows.item(0).seq;
-            var sql = 'UPDATE ' + BY_SEQ_STORE +
-              ' SET json=?, deleted=? WHERE doc_id=? AND rev=?;';
-            var sqlArgs = [json, deletedInt, id, rev];
-            tx.executeSql(sql, sqlArgs, function (tx) {
-              insertAttachmentMappings(seq, function () {
-                dataWritten(tx, seq);
-              });
-            });
-          });
-          return false; // ack that we've handled the error
-        });
-      }
-
-      function collectResults(attachmentErr) {
-        if (!err) {
-          if (attachmentErr) {
-            err = attachmentErr;
-            callback(err);
-          } else if (recv === attachments.length) {
-            finish();
-          }
-        }
-      }
-
-      var err = null;
-      var recv = 0;
-
-      docInfo.data._id = docInfo.metadata.id;
-      docInfo.data._rev = docInfo.metadata.rev;
-      var attachments = Object.keys(docInfo.data._attachments || {});
-
-
-      if (deleted) {
-        docInfo.data._deleted = true;
-      }
-
-      function attachmentSaved(err) {
-        recv++;
-        collectResults(err);
-      }
-
-      attachments.forEach(function (key) {
-        var att = docInfo.data._attachments[key];
-        if (!att.stub) {
-          var data = att.data;
-          delete att.data;
-          var digest = att.digest;
-          saveAttachment(digest, data, attachmentSaved);
-        } else {
-          recv++;
-          collectResults();
-        }
-      });
-
-      if (!attachments.length) {
-        finish();
-      }
-
-      function autoCompact() {
-        if (!isUpdate || !api.auto_compaction) {
-          return; // nothing to do
-        }
-        var id = docInfo.metadata.id;
-        var revsToDelete = utils.compactTree(docInfo.metadata);
-        compactRevs(revsToDelete, id, tx);
-      }
-
-      function dataWritten(tx, seq) {
-        autoCompact();
-        docInfo.metadata.seq = seq;
-        delete docInfo.metadata.rev;
-
-        var sql = isUpdate ?
-          'UPDATE ' + DOC_STORE +
-          ' SET json=?, max_seq=?, winningseq=' +
-          '(SELECT seq FROM ' + BY_SEQ_STORE +
-          ' WHERE doc_id=' + DOC_STORE + '.id AND rev=?) WHERE id=?'
-          : 'INSERT INTO ' + DOC_STORE +
-          ' (id, winningseq, max_seq, json) VALUES (?,?,?,?);';
-        var metadataStr = utils.safeJsonStringify(docInfo.metadata);
-        var id = docInfo.metadata.id;
-        var params = isUpdate ?
-          [metadataStr, seq, winningRev, id] :
-          [id, seq, seq, metadataStr];
-        tx.executeSql(sql, params, function () {
-          results[resultsIdx] = docInfo;
-          fetchedDocs.set(id, docInfo.metadata);
-          callback();
-        });
-      }
-    }
-
-    function processDocs() {
-      utils.processDocs(docInfos, api, fetchedDocs,
-        tx, results, writeDoc, opts);
-    }
-
-    function fetchExistingDocs(callback) {
-      if (!docInfos.length) {
-        return callback();
-      }
-
-      var numFetched = 0;
-
-      function checkDone() {
-        if (++numFetched === docInfos.length) {
-          callback();
-        }
-      }
-
-      docInfos.forEach(function (docInfo) {
-        if (docInfo._id && utils.isLocalId(docInfo._id)) {
-          return checkDone(); // skip local docs
-        }
-        var id = docInfo.metadata.id;
-        tx.executeSql('SELECT json FROM ' + DOC_STORE +
-          ' WHERE id = ?', [id], function (tx, result) {
-          if (result.rows.length) {
-            var metadata = utils.safeJsonParse(result.rows.item(0).json);
-            fetchedDocs.set(id, metadata);
-          }
-          checkDone();
-        });
-      });
-    }
-
-    function saveAttachment(digest, data, callback) {
-      var sql = 'SELECT digest FROM ' + ATTACH_STORE + ' WHERE digest=?';
-      tx.executeSql(sql, [digest], function (tx, result) {
-        if (result.rows.length) { // attachment already exists
-          return callback();
-        }
-        // we could just insert before selecting and catch the error,
-        // but my hunch is that it's cheaper not to serialize the blob
-        // from JS to C if we don't have to (TODO: confirm this)
-        sql = 'INSERT INTO ' + ATTACH_STORE +
-          ' (digest, body, escaped) VALUES (?,?,1)';
-        tx.executeSql(sql, [digest, escapeBlob(data)], function () {
-          callback();
-        }, function () {
-          // ignore constaint errors, means it already exists
-          callback();
-          return false; // ack we handled the error
-        });
-      });
-    }
-
-    utils.preprocessAttachments(docInfos, 'binary', function (err) {
-      if (err) {
-        return callback(err);
-      }
-      db.transaction(function (txn) {
-        tx = txn;
-        verifyAttachments(function (err) {
-          if (err) {
-            preconditionErrored = err;
-          } else {
-            fetchExistingDocs(processDocs);
-          }
-        });
-      }, unknownError(callback), complete);
-    });
+    websqlBulkDocs(req, opts, api, db, WebSqlPouch.Changes, callback);
   };
 
   api._get = function (id, opts, callback) {
@@ -1078,8 +596,8 @@ function WebSqlPouch(opts, callback) {
 
   function countDocs(tx, callback) {
 
-    if (docCount !== -1) {
-      return callback(docCount);
+    if (api._docCount !== -1) {
+      return callback(api._docCount);
     }
 
     // count the total rows
@@ -1090,8 +608,8 @@ function WebSqlPouch(opts, callback) {
       BY_SEQ_STORE + '.deleted=0');
 
     tx.executeSql(sql, [], function (tx, result) {
-      docCount = result.rows.item(0).num;
-      callback(docCount);
+      api._docCount = result.rows.item(0).num;
+      callback(api._docCount);
     });
   }
 
@@ -1161,11 +679,12 @@ function WebSqlPouch(opts, callback) {
           for (var i = 0, l = result.rows.length; i < l; i++) {
             var item = result.rows.item(i);
             var metadata = utils.safeJsonParse(item.metadata);
-            var data = unstringifyDoc(item.data, metadata.id, item.rev);
+            var id = metadata.id;
+            var data = unstringifyDoc(item.data, id, item.rev);
             var winningRev = data._rev;
             var doc = {
-              id: metadata.id,
-              key: metadata.id,
+              id: id,
+              key: id,
               value: {rev: winningRev}
             };
             if (opts.include_docs) {
@@ -1201,12 +720,12 @@ function WebSqlPouch(opts, callback) {
     opts = utils.clone(opts);
 
     if (opts.continuous) {
-      var id = name + ':' + utils.uuid();
-      WebSqlPouch.Changes.addListener(name, id, api, opts);
-      WebSqlPouch.Changes.notify(name);
+      var id = api._name + ':' + utils.uuid();
+      WebSqlPouch.Changes.addListener(api._name, id, api, opts);
+      WebSqlPouch.Changes.notify(api._name);
       return {
         cancel: function () {
-          WebSqlPouch.Changes.removeListener(name, id);
+          WebSqlPouch.Changes.removeListener(api._name, id);
         }
       };
     }
@@ -1329,7 +848,7 @@ function WebSqlPouch(opts, callback) {
       // but that's not performant. after migration 6, we remove \u0000
       // and add it back in afterwards
       var item = result.rows.item(0);
-      var data = item.escaped ? unescapeBlob(item.body) :
+      var data = item.escaped ? websqlUtils.unescapeBlob(item.body) :
         parseHexString(item.body, encoding);
       if (opts.encode) {
         res = btoa(data);
