+++ lib/index.js
@@ -11,12 +11,12 @@ PouchDB.Errors = require('./deps/errors');
 PouchDB.replicate = require('./replicate').replicate;
 PouchDB.sync = require('./sync');
 PouchDB.version = require('./version');
-var httpAdapter = require('./adapters/http');
+var httpAdapter = require('./adapters/http/http');
 PouchDB.adapter('http', httpAdapter);
 PouchDB.adapter('https', httpAdapter);
 
 PouchDB.adapter('idb', require('./adapters/idb/idb'));
-PouchDB.adapter('websql', require('./adapters/websql'));
+PouchDB.adapter('websql', require('./adapters/websql/websql'));
 PouchDB.plugin(require('pouchdb-mapreduce'));
 
 if (!process.browser) {
