+++ lib/utils.js
@@ -762,7 +762,7 @@ exports.processDocs = function processDocs(docInfos, api, fetchedDocs,
           if (err) {
             results[resultsIdx] = err;
           } else {
-            results[resultsIdx] = {};
+            results[resultsIdx] = {ok: true};
           }
           checkAllDocsDone();
         });
